LPS-137441 - Move file to be in same location as the interface they implement
