+++ modules/apps/frontend-icons/frontend-icons-web/src/main/java/com/liferay/frontend/icons/web/internal/configuration/admin/category/FrontendIconsConfigurationCategory.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.frontend.icons.web.internal.portal.settings.configuration.admin.category;
+package com.liferay.frontend.icons.web.internal.configuration.admin.category;
 
 import com.liferay.configuration.admin.category.ConfigurationCategory;
