+++ modules/apps/frontend-icons/frontend-icons-web/src/main/java/com/liferay/frontend/icons/web/internal/portlet/action/SaveFrontendIconsPackFromExistingIconsMVCActionCommand.java
@@ -79,7 +79,8 @@ public class SaveFrontendIconsPackFromExistingIconsMVCActionCommand
 			return;
 		}
 
-		String iconPackName = ParamUtil.getString(actionRequest, "iconPackName");
+		String iconPackName = ParamUtil.getString(
+			actionRequest, "iconPackName");
 
 		String icons = ParamUtil.getString(actionRequest, "icons");
